Prepare command add_workshop_from_ods for Period 2
Management command add_workshop_from_ods now takes a second argument
period_n in order to choose the period number that is going to be
lined with the new workshopperiods created. The selected period must
be created manually (stop execution if period doesn't exists).

The shedules lookup by `datetime.time` failed localy but not in
production. This could be due to differences in the way mysql and
sqlite store time. To fix this, `time__hour` and `time__minute` lookup
is used so it compares ints (doesn't fail localy).

Also some fixes:

  - Cristina and Javiera username dict entries
  - "Mañíos" cycle name with tilde